0.1 version :
First working version, basic (very basic for now) implementation of AVFoundation framework using Swift 3.
As of this release is implemented the AVCaptureDeviceDiscovery part, so front and back cameras are assigned to their AVCaptureDevice and an AVCaptureInput is instantiated using the AVCaptureDevice that represents the (wide) backcamera.
Also a very basic istance of AVCaptureSession is run to test if this AVCaptureInput can be added as input for an AVCaptureSession.

Everything is shown as a layer (AVCapturePreviewLayer) over the default ViewController's view.  Basic constrains are setted up via Storyboard's controls but in the future I'll try to implement this via code. Those constrains, as of now, are working only in Landscape mode.

TODO:
- add func to take photo to the circular button
- set constrains for landscape
-